I hate lots of code in one file, so I'm splitting the imgur upload code into its own module. This will make the app.js file cleaner and easier to read.
