+++ src/app.js
@@ -6,9 +6,8 @@ import fs from 'fs';
 import mime from 'mime';
 import QRCode from 'qrcode';
 import ExifReader from 'exifreader';
-import axios from 'axios';
-import FormData from 'form-data';
 import { spawn } from 'node:child_process';
+import { uploadToImgur } from './modules/imgur-upload.js';
 
 // Initialize Express
 const app = express();
@@ -150,6 +149,7 @@ function sendAnalytics(userAgent, clientIp, data) {
     body: JSON.stringify(data),
   });
 }
+export {sendAnalytics};
 
 // Function to detect software title from image EXIF data or file name
 async function getSoftwareTitle(imgFile, mimeType, originalFileName) {
@@ -405,11 +405,11 @@ app.post(['/', '/m', '/m/'], upload.single('img'), async function (req, res, err
         res.end(renderMessage(String(req.get('User-Agent')), connectedHost, req.path.startsWith('/m'), 'The file you uploaded is not supported by Imgur. Please select a PNG or JPEG image.'));
         return;
       } else {
-        uploadResult = await uploadToImgur(req.file, softwareTitle);
+        uploadResult = await uploadToImgur(req.file, softwareTitle, imgurClientId, plausibleDomain, req);
       }
     } else {
       req.body['upload-type'] === 'imageshare'
-      uploadResult = await uploadToLocal(req.file, protocol, connectedHost);
+      uploadResult = await uploadToLocal(req.file, protocol, connectedHost, req);
     }
 
     if (uploadResult.success) {
@@ -432,7 +432,7 @@ app.post(['/', '/m', '/m/'], upload.single('img'), async function (req, res, err
 });
 
 // Function to handle local uploads
-async function uploadToLocal(filePath, protocol, connectedHost) {
+async function uploadToLocal(filePath, protocol, connectedHost, req) {
   return new Promise((resolve) => {
     // Create unique shortlink that isn't already in storage
     let shortLink;
@@ -467,93 +467,6 @@ async function uploadToLocal(filePath, protocol, connectedHost) {
     }
     // Return success and desplay results to user :3
     resolve({ success: true, link: `${protocol}://${connectedHost}/i/${shortLink}`, qrLink: `${protocol}://${connectedHost}/${filePath.path.replace('uploads/', 'qr/')}` });
-  });
-}
-
-// Function to handle Imgur uploads
-async function uploadToImgur(filePath, softwareTitle) {
-  // Send Imgur upload request with the file
-  return new Promise((resolve) => {
-    //Make sure that the current time stamp is past the rate limit reset time
-    let rateLimitReset = 0;
-    if (fs.existsSync('rateLimitReset.txt')) {
-      rateLimitReset = fs.readFileSync('rateLimitReset.txt', 'utf8');
-    }
-    if (rateLimitReset > Math.floor(Date.now() / 1000)) {
-      // Error to move to catch block and delete the file
-      // Remove cached file and return failure
-      fs.unlinkSync(filePath.path);
-      console.log('Prevented upload to imgur')
-      console.log(`Deleted cached file: ${filePath.path}`);
-      console.log(`There was an error uploading to Imgur. Make sure that your API Key is set and you haven't exceeded your rate limit.`);
-      resolve({ success: false, reason: `Imgur is currently at max capacity, so you will have to try again later. You can still upload to ImageShare though!` });
-      return;
-    }
-
-
-    // Set up Imgur API request
-    const formData = new FormData();
-    formData.append('image', fs.createReadStream(filePath.path));
-    formData.append('type', 'file');
-    formData.append('title', `${softwareTitle}`);
-    formData.append('description', 'Uploaded by ImageShare (github -> corbindavenport/imageshare)');
-
-    let options = {
-      method: 'post',
-      maxBodyLength: Infinity,
-      url: 'https://api.imgur.com/3/image',
-      headers: {
-        'Authorization': `Client-ID ${imgurClientId}`,
-        ...formData.getHeaders()
-      },
-      data: formData
-    };
-
-    // Send Results after we get api results
-    axios.request(options)
-      // Get header response for x-post-rate-limit-remaining
-      .then(response => {
-        console.log(`Imgur API Rate Limit Remaining: ${response.headers['x-post-rate-limit-remaining']}`);
-        if (response.headers['x-post-rate-limit-remaining'] < 10) {
-          console.log(`Imgur API Rate Limit Remaining is below 10. Please check your rate limit.`);
-          console.log(`You must wait: ${response.headers['x-post-rate-limit-reset']} seconds before you can upload again.`);
-
-          // Get unix timestamp for when the rate limit will be reset
-          const restTime = (Number(Math.floor(Date.now() / 1000)) + Number(response.headers['x-post-rate-limit-reset']));
-          console.log(`UnixTimestamp: ${restTime}`)
-
-          // Create a file with the rate limit reset time (overwrite if it exists already)
-          fs.writeFileSync('rateLimitReset.txt', restTime.toString());
-        }
-        return response.data;
-      })
-      .then(body => {
-        // Check if upload was successful and handle accordingly        
-        if (body.success) {
-          //Remove cached file
-          fs.unlinkSync(filePath.path);
-          console.log(`Deleted cached file: ${filePath.path}`);
-          // Return success and link
-          resolve({ success: true, link: `https://imgur.com/${body.data.id}`, qrLink: `/qr/${body.data.link}` });
-        }
-
-        // Send async Plausible analytics page view if enabled
-        if (plausibleDomain) {
-          const data = {
-            name: 'Upload',
-            props: JSON.stringify({ 'Upload Mode': 'Native' }),
-            url: '/',
-            domain: plausibleDomain
-          };
-          sendAnalytics(req.get('User-Agent'), (req.headers['x-forwarded-for'] || req.ip), data);
-        }
-      })
-      .catch(error => {
-        // Remove cached file and return failure
-        fs.unlinkSync(filePath.path);
-        console.log(`Deleted cached file: ${filePath.path}`);
-        resolve({ success: false, reason: `There was an error uploading to Imgur. Make sure that your API Key is set and you haven't exceeded your rate limit.` });
-      });
   });
 }
